fixed one bug; some acceleration
